feat(ownership): fast per-file owner resolution + top-of-file annotation support
