+++ src/ownership.rs
@@ -14,6 +14,7 @@ mod file_owner_finder;
 pub(crate) mod mapper;
 pub(crate) mod parser;
 mod validator;
+pub(crate) mod fast;
 
 use crate::{
     ownership::mapper::DirectoryMapper,
